# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "laravel-api-main",
        "description": "Description de votre API",
        "contact": {
            "name": "Nom du contact",
            "email": "contact@example.com"
        },
        "license": {
            "name": "Licence",
            "url": "https://example.com/license"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Liste des produits",
                "operationId": "7ca3593d05653d957ae7be1ac648ca44",
                "responses": {
                    "200": {
                        "description": "Retourne la liste des produits",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Ajouter un nouveau produit",
                "operationId": "cfdb2a3921bee30f9cad135e1c5f1452",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Product"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Retourne le produit ajouté",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Afficher un produit spécifique",
                "operationId": "1948ab0c6b14a722ad6d1665cd38e1f1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du produit",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retourne le produit spécifique",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Mettre à jour un produit",
                "operationId": "b9e4a8996649edabc6570985e351fa6b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du produit à mettre à jour",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Product"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Retourne le produit mis à jour",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Supprimer un produit",
                "operationId": "c935132359292f5732d896b035f3f322",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du produit à supprimer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Message de confirmation de suppression"
                    }
                }
            }
        },
        "/api/v1/products/{id}/edit": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Afficher le formulaire d'édition d'un produit",
                "operationId": "2a70ae42cf6a0be5d90088ce02946e78",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du produit à éditer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retourne le produit à éditer",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Categorie": {
                "title": "Categorie",
                "description": "Categorie model",
                "properties": {
                    "id": {
                        "description": "Categorie ID",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Categorie name",
                        "type": "string"
                    },
                    "description": {
                        "description": "Categorie description",
                        "type": "string"
                    },
                    "products": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Product"
                        }
                    },
                    "created_at": {
                        "description": "Date and time when the categorie was created",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Date and time when the categorie was last updated",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Product": {
                "title": "Product",
                "description": "Product model",
                "properties": {
                    "id": {
                        "description": "Product ID",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Product name",
                        "type": "string"
                    },
                    "price": {
                        "description": "Product price",
                        "type": "number",
                        "format": "float"
                    },
                    "stock": {
                        "description": "Product stock",
                        "type": "integer"
                    },
                    "description": {
                        "description": "Product description",
                        "type": "string"
                    },
                    "image": {
                        "description": "URL of the product image",
                        "type": "string"
                    },
                    "categories": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Categorie"
                        }
                    },
                    "created_at": {
                        "description": "Date and time when the product was created",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Date and time when the product was last updated",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    }
}